인증, 인가 기능 구현

+++ src/main/java/com/sparta/Jmt/service/PostService.java
@@ -4,11 +4,14 @@ import com.sparta.Jmt.dto.PostListResponseDto;
 import com.sparta.Jmt.dto.PostRequestDto;
 import com.sparta.Jmt.dto.PostResponseDto;
 import com.sparta.Jmt.entity.Post;
+import com.sparta.Jmt.entity.User;
 import com.sparta.Jmt.repository.PostRepository;
-import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.concurrent.RejectedExecutionException;
+import java.util.stream.Collectors;
 
 @Service
 public class PostService {
@@ -19,51 +22,58 @@ public class PostService {
         this.postRepository = postRepository;
     }
 
-    public PostResponseDto createPost(PostRequestDto postRequestDto){
+    public PostResponseDto createPost(PostRequestDto postRequestDto, User user){
+        // RequestDto -> Entity(게시글 생성)
         Post post = new Post(postRequestDto);
+        post.setUser(user);
+        // DB 저장
         postRepository.save(post);
-
-        PostResponseDto postResponseDto = new PostResponseDto(post);
-        return postResponseDto;
+        return new PostResponseDto(post);
     }
 
     //전체 게시물 보기
     public PostListResponseDto getPosts() {
-        PostListResponseDto postListResponseDto = new PostListResponseDto(postRepository.findAllByOrderByModifiedAtDesc().stream().map(PostResponseDto::new).toList());
-
-        return postListResponseDto;
+        List<PostResponseDto> postList = postRepository.findAll().stream().map(PostResponseDto::new).collect(Collectors.toList());
+        return new PostListResponseDto(postList);
     }
 
     //선택한 게시물 보기
-    public PostResponseDto getPost(Long postId) {
+    public PostResponseDto getPostById(Long postId) {
         //Post형태로 postId에 맞는 게시글 찾기
-        Post post = findById(postId);
+        Post post = findPost(postId);
 
-        //Post 형태의 찾은 게시물을 ResponseDto로 바꿔주기
-        PostResponseDto postResponseDto = new PostResponseDto(post);
-        return postResponseDto;
+        //Post 형태의 찾은 게시물을 ResponseDto로 받아오기
+        return new PostResponseDto(post);
     }
 
-    public void updatePost(Long postId,PostRequestDto requestDto){
-        //Post형태로 postId에 맞는 게시글 찾기
-        Post post = findById(postId);
+    @Transactional
+    public PostResponseDto updatePost(Long postId,PostRequestDto requestDto, User user){
+        // 해당 포스트가 DB에 존재하는지 확인
+        Post post = findPost(postId);
+        //게시글 작성자(post.user)와 요청자(user)가 같은지 학인
+        if(!post.getUser().equals(user)) {
+            throw new RejectedExecutionException();
+        }
+        // post 내용 수정
+        post.setPostTitle(requestDto.getPostTitle());
+        post.setPostContent(requestDto.getPostContent());
 
-        //불러온 게시물을 수정
-        post.updatePost(requestDto);
+        return new PostResponseDto(post);
     }
-
-    public void deletePost(Long postId){
-        //Post형태로 postId에 맞는 게시글 찾기
-        Post post = findById(postId);
-
-        //불러온 게시물을 삭제
+    public void deletePost(Long postId, User user){
+        // 해당 포스트가 DB에 존재하는지 확인
+        Post post = findPost(postId);
+        //게시글 작성자(post.user)와 요청자(user)가 같은지 학인
+        if(!post.getUser().equals(user)) {
+            throw new RejectedExecutionException();
+        }
+        // post 삭제
         postRepository.delete(post);
     }
 
-    public Post findById(Long postId){
+    private Post findPost(Long postId) {
         return postRepository.findById(postId).orElseThrow(() ->
                 new IllegalArgumentException("선택한 게시글이 존재하지 않습니다."));
     }
-
 
 }
